% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bg_transpose}
\alias{bg_transpose}
\title{bg_transpose}
\usage{
bg_transpose(mat, byrow = FALSE)
}
\arguments{
\item{mat}{A Matrix.}

\item{byrow}{\code{FALSE} computes based on row index. \code{TRUE} computes based on column index.}
}
\value{
A matrix transposed.
}
\description{
Transpose of a matrix based on row or column index.
}
