% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distance_weighted_gauss}
\alias{distance_weighted_gauss}
\title{distance_weighted_gauss}
\usage{
distance_weighted_gauss(coord_xy, point_xy, tau)
}
\arguments{
\item{coord_xy}{A Matrix with lon/lat coordinates.}

\item{point_xy}{lon/lat coordinate.}

\item{tau}{bandwidth.}
}
\value{
A vector.
}
\description{
Weighted Distance based on Gaussian function
}
