% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{float_relative}
\alias{float_relative}
\title{float_relative}
\usage{
float_relative(xx)
}
\arguments{
\item{xx}{A Matrix with lon/lat coordinates.}
}
\value{
A vector.
}
\description{
Relative sum formula
}
