% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{float_round}
\alias{float_round}
\title{float_round}
\usage{
float_round(float_n, digits = 0L)
}
\arguments{
\item{float_n}{A numeric vector.}

\item{digits}{integer indicating the number of decimal places .}
}
\value{
A vector.
}
\description{
Rouding of Numbers
}
