% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matrixcec_square}
\alias{matrixcec_square}
\title{matrixcec_square}
\usage{
matrixcec_square(X, y)
}
\arguments{
\item{X}{A Matrix.}

\item{y}{A Vector}
}
\value{
A matrix.
}
\description{
Matrix Square
}
