% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matrixvec_plus}
\alias{matrixvec_plus}
\title{matrixvec_plus}
\usage{
matrixvec_plus(X, y)
}
\arguments{
\item{X}{A Matrix.}

\item{y}{A Vector}
}
\value{
A matrix.
}
\description{
Matrix Multiplication
}
