% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matrixvec_subs}
\alias{matrixvec_subs}
\title{matrixvec_subs}
\usage{
matrixvec_subs(X, y)
}
\arguments{
\item{X}{A Matrix.}

\item{y}{A Vector}
}
\value{
A matrix.
}
\description{
Matrix Substraction
}
