% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invisible.R
\name{plot,SpatHetero,ANY-method}
\alias{plot,SpatHetero,ANY-method}
\title{Plot Heterogeneity Layer}
\usage{
\S4method{plot}{SpatHetero,ANY}(x, comp = NULL, ...)
}
\arguments{
\item{x}{SpatHetero Class}

\item{comp}{integer. Plot specific component of the heterogeneity.}

\item{...}{Plot parameters forwarded.}
}
\value{
No return value, called for side effects.
}
\description{
Plot
}
