% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invisible.R
\docType{class}
\name{SpatHetero-class}
\alias{SpatHetero-class}
\title{SpatHetero}
\description{
SpatHetero
}
\section{Slots}{

\describe{
\item{\code{hetero}}{A Heterogeneity Layer}

\item{\code{matrix}}{SpatHetero_in data}

\item{\code{rasters}}{A SpatRaster for Each Component}
}}

